# internal/api/openapi.yaml
openapi: "3.1.0"
info:
  title: sqi-server API
  description: |
    REST API for the sqi distributed task management server.

    All error responses use the RFC 7807 problem-details format
    (`application/problem+json`).  Clients should branch on `Content-Type`
    rather than status code alone to detect error bodies.
  version: "0.1.0"
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /api/v1
    description: Current server
...
components:
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource UUID
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 50
      description: Maximum number of items to return (1–1000).
    offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Zero-based item offset for pagination.
    sortDir:
      name: sort_dir
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
      description: Sort direction.
  schemas:
...
    ProblemDetail:
      type: object
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          example: "about:blank"
        title:
          type: string
          example: Not Found
        status:
          type: integer
          example: 404
        detail:
          type: string
          example: job not found
        instance:
          type: string
          description: Request-ID that identifies this specific occurrence.
          example: a1b2c3d4e5f60708

    # ── Jobs ─────────────────────────────────────────────────────────────────
...
    GPUInfo:
      type: object
      properties:
        vendor:
          type: string
        model:
          type: string
        vram_mb:
          type: integer
        count:
          type: integer
    Worker:
      type: object
      required: [id, farm_id, hostname, gpu, status, registered_at, updated_at]
      properties:
        id:
          type: string
        farm_id:
          type: string
        queue_id:
          type: string
        hostname:
          type: string
        ip_address:
          type: string
        compute_location:
          type: string
          description: Named compute location (e.g. "on-prem", "aws-us-east-1").
        os:
          type: string
        os_version:
          type: string
        cpu_count:
          type: integer
        ram_mb:
          type: integer
        gpu:
          $ref: "#/components/schemas/GPUInfo"
        tags:
          type: object
          additionalProperties:
            type: string
          description: Arbitrary capability tags used for task matching.
        status:
          type: string
          enum: [online, offline, disabled]
        last_heartbeat_at:
          type: string
          format: date-time
          nullable: true
        registered_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CurrentTask:
      type: object
      required: [id, job_id, name, status]
      properties:
        id:
          type: string
        job_id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [assigned, running]
        assigned_at:
          type: string
          format: date-time
          nullable: true
    WorkerDetail:
      allOf:
        - $ref: "#/components/schemas/Worker"
        - type: object
          properties:
            current_task:
              $ref: "#/components/schemas/CurrentTask"
              nullable: true
    WorkerList:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Worker"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    WorkerActionResponse:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
        status:
          type: string
          enum: [online, offline, disabled]

    # ── Farms ─────────────────────────────────────────────────────────────────
...
  responses:
...
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
    InternalServerError:
      description: Internal server error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"

# ── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /workers:
    get:
      operationId: listWorkers
      summary: List workers
      tags: [workers]
      parameters:
        - name: farm_id
          in: query
          schema:
            type: string
        - name: queue_id
          in: query
          schema:
            type: string
        - name: compute_location
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [online, offline, disabled]
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [hostname, status, registered_at, last_heartbeat_at]
            default: hostname
        - $ref: "#/components/parameters/sortDir"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: Paginated list of workers.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerList"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /workers/{id}:
    parameters:
      - $ref: "#/components/parameters/id"

    get:
      operationId: getWorker
      summary: Get a worker
      tags: [workers]
      description: Returns worker detail including capabilities and the task currently being executed, if any.
      responses:
        "200":
          description: Worker detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /workers/{id}/disable:
    parameters:
      - $ref: "#/components/parameters/id"

    post:
      operationId: disableWorker
      summary: Disable a worker
      tags: [workers]
      description: |
        Administratively pauses the worker so it receives no new task
        assignments.  Workers already in `disabled` state are accepted
        idempotently.
      responses:
        "200":
          description: Worker status after the operation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerActionResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /workers/{id}/enable:
    parameters:
      - $ref: "#/components/parameters/id"

    post:
      operationId: enableWorker
      summary: Enable a worker
      tags: [workers]
      description: |
        Re-enables a disabled worker.  Online or offline workers are accepted
        idempotently; their status is returned unchanged.
      responses:
        "200":
          description: Worker status after the operation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerActionResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ── Farms ──────────────────────────────────────────────────────────────────
...